+++ tests/constants.py
@@ -1,0 +1,13 @@
+LARGE_PARAGRAPH = """I know this from my own experience, as a high school kid
+writing programs in Basic. That language didn't even support recursion. It's
+hard to imagine writing programs without using recursion, but I didn't miss it
+at the time. I thought in Basic. And I was a whiz at it.
+Master of all I surveyed. - Paul Graham
+
+A programming language is for thinking about programs, not for expressing
+programs you've already thought of. It should be a pencil, not a pen.
+- Paul Graham
+
+A recursive definition does not necessarily lead to a recursive process.
+- Gerald Jay Sussman
+"""
